The number of items a player can have is now bounded
It is 4, and if a player has already 4 items, then the zombie doesn't try to carry the item.
If a zombie is already carrying an item and fetches it to the fortress but the player cannot take anymore item, then it is dropped here and the zombie continues with its life... Or death... I don't know...
For issue 6.