Cities' number of inhabitants is represented by a side-bar
Now, a bar represents the number of inhabitants in a city.
The height represent the maximum number of inhabitants while the filling represent the current count.
The filling color is either gray, red or blue, depending on the player controlling it.
The sprites have been redone to make some free space for the bar to be displayed properly.
For issue 14.